0.13.0
- added `Tier` enum
- changed a lot of unsigned ints to signed ones because apparrently the
  API likes to return stupid values
- changed `map_name` on `Map` to an Option because fuck you that's why

+++ src/global_api/maps/mod.rs
@@ -26,11 +26,11 @@ pub async fn get(params: Params, client: &crate::Client) -> Result<Vec<Map>, Err
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct Map {
-	pub id: u32,
+	pub id: i32,
 	pub name: String,
-	pub filesize: u64,
+	pub filesize: i64,
 	pub validated: bool,
-	pub difficulty: u8,
+	pub difficulty: i32,
 	pub created_on: String,
 	pub updated_on: String,
 	pub approved_by_steamid64: String,
@@ -42,12 +42,12 @@ api_response!(Map);
 
 #[derive(Default, Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct Params {
-	pub id: Option<u32>,
+	pub id: Option<i32>,
 	pub name: Option<String>,
-	pub larger_than_filesize: Option<u32>,
-	pub smaller_than_filesize: Option<u32>,
+	pub larger_than_filesize: Option<i32>,
+	pub smaller_than_filesize: Option<i32>,
 	pub is_validated: Option<bool>,
-	pub difficulty: Option<u8>,
+	pub difficulty: Option<i32>,
 	pub created_since: Option<String>,
 	pub updated_since: Option<String>,
 	pub offset: Option<i32>,
