+++ src/global_api/record_filters/mod.rs
@@ -23,11 +23,11 @@ pub async fn get(params: Params, client: &crate::Client) -> Result<Vec<RecordFil
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct RecordFilter {
-	pub id: u32,
-	pub map_id: u32,
-	pub stage: u8,
-	pub mode_id: u8,
-	pub tickrate: u8,
+	pub id: i32,
+	pub map_id: i32,
+	pub stage: i32,
+	pub mode_id: i32,
+	pub tickrate: i32,
 	pub has_teleports: bool,
 	pub created_on: String,
 	pub updated_by_id: String,
@@ -37,11 +37,11 @@ api_response!(RecordFilter);
 
 #[derive(Default, Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct Params {
-	pub ids: Option<u32>,
-	pub map_ids: Option<u32>,
-	pub stages: Option<u8>,
-	pub mode_ids: Option<u8>,
-	pub tickrates: Option<u8>,
+	pub ids: Option<i32>,
+	pub map_ids: Option<i32>,
+	pub stages: Option<i32>,
+	pub mode_ids: Option<i32>,
+	pub tickrates: Option<i32>,
 	pub has_teleports: Option<bool>,
 	pub offset: Option<i32>,
 	pub limit: Option<u32>,
